Adding Auth module (#42)
* Migrate `impl AuthQuery for BaseCoinApp` to `app/modules/auth.rs`

* Fix imports.
