+++ src/app/mod.rs
@@ -9,11 +9,7 @@ use crate::app::response::ResponseFromErrorExt;
 use crate::app::store::{
     Height, Identifier, Path, ProvableStore, RevertibleStore, SharedStore, Store, SubStore,
 };
-use crate::prostgen::cosmos::auth::v1beta1::{
-    query_server::Query as AuthQuery, BaseAccount, QueryAccountRequest, QueryAccountResponse,
-    QueryAccountsRequest, QueryAccountsResponse, QueryParamsRequest as AuthQueryParamsRequest,
-    QueryParamsResponse as AuthQueryParamsResponse,
-};
+use crate::prostgen::cosmos::auth::v1beta1::BaseAccount;
 use crate::prostgen::cosmos::base::tendermint::v1beta1::{
     service_server::Service as HealthService, GetBlockByHeightRequest, GetBlockByHeightResponse,
     GetLatestBlockRequest, GetLatestBlockResponse, GetLatestValidatorSetRequest,
@@ -347,42 +343,6 @@ impl<S: ProvableStore + 'static> HealthService for BaseCoinApp<S> {
         &self,
         _request: Request<GetValidatorSetByHeightRequest>,
     ) -> Result<Response<GetValidatorSetByHeightResponse>, Status> {
-        unimplemented!()
-    }
-}
-
-#[tonic::async_trait]
-impl<S: ProvableStore + 'static> AuthQuery for BaseCoinApp<S> {
-    async fn accounts(
-        &self,
-        _request: Request<QueryAccountsRequest>,
-    ) -> Result<Response<QueryAccountsResponse>, Status> {
-        unimplemented!()
-    }
-
-    async fn account(
-        &self,
-        _request: Request<QueryAccountRequest>,
-    ) -> Result<Response<QueryAccountResponse>, Status> {
-        debug!("Got auth account request");
-
-        let mut account = self.account.write().unwrap();
-        let mut buf = Vec::new();
-        account.encode(&mut buf).unwrap(); // safety - cannot fail since buf is a vector
-        account.sequence += 1;
-
-        Ok(Response::new(QueryAccountResponse {
-            account: Some(Any {
-                type_url: "/cosmos.auth.v1beta1.BaseAccount".to_string(),
-                value: buf,
-            }),
-        }))
-    }
-
-    async fn params(
-        &self,
-        _request: Request<AuthQueryParamsRequest>,
-    ) -> Result<Response<AuthQueryParamsResponse>, Status> {
         unimplemented!()
     }
 }
